+++ src/config.ts
@@ -7,3 +7,5 @@ export const FFMPEG_RELEASES_API =
   'https://api.github.com/repos/jellyfin/jellyfin-ffmpeg/releases?per_page=20';
 export const RSGAIN_RELEASE_API =
   'https://api.github.com/repos/complexlogic/rsgain/releases/tags/v3.7';
+export const CCEXTRACTOR_RELEASE_API =
+  'https://api.github.com/repos/CCExtractor/ccextractor/releases/latest';